+++ src/helpers/sqs.rs
@@ -22,9 +22,9 @@ pub struct SQSMessage {
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct SMPCRequest {
-    pub request_id:   String,
-    pub iris_code:    String,
-    pub mask_code:    String,
+    pub request_id: String,
+    pub iris_code:  String,
+    pub mask_code:  String,
 }
 
 impl SMPCRequest {
@@ -44,19 +44,14 @@ impl SMPCRequest {
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct ResultEvent {
-    pub node_id:   usize,
-    pub db_index:  u32,
-    pub is_match:  bool,
+    pub node_id:    usize,
+    pub db_index:   u32,
+    pub is_match:   bool,
     pub request_id: String,
 }
 
 impl ResultEvent {
-    pub fn new(
-        node_id: usize,
-        db_index: u32,
-        is_match: bool,
-        request_id: String,
-    ) -> Self {
+    pub fn new(node_id: usize, db_index: u32, is_match: bool, request_id: String) -> Self {
         Self {
             node_id,
             db_index,
